Fixlint by separate RootView file for macOS
